# docs/swagger.yaml
definitions:
  api.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
...
  requests.LoginRequest:
    properties:
      email:
        example: username@gmail.com
        type: string
      password:
        example: P@ssw0r6
        type: string
    required:
    - email
    - password
    type: object
  requests.RefreshRequest:
    properties:
      token:
        example: refresh_token
        type: string
    required:
    - token
    type: object
...
  responses.LoginResponse:
    properties:
      accessToken:
        type: string
      exp:
        type: integer
      refreshToken:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/domain:
    get:
      consumes:
      - application/json
      description: Returns a sample response for listing domains.
      operationId: domain-list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: List domains
      tags:
      - Domain Management
    post:
      consumes:
      - application/json
      description: Not implemented.
      operationId: domain-create
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Create domain
      tags:
      - Domain Management
  /api/domain/{uuid}:
    delete:
      consumes:
      - application/json
      description: Not implemented.
      operationId: domain-delete
      parameters:
      - description: Domain UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete domain
      tags:
      - Domain Management
    get:
      consumes:
      - application/json
      description: Not implemented.
      operationId: domain-read
      parameters:
      - description: Domain UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Get domain
      tags:
      - Domain Management
    put:
      consumes:
      - application/json
      description: Not implemented.
      operationId: domain-update
      parameters:
      - description: Domain UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Update domain
      tags:
      - Domain Management
...
  /login:
    post:
      consumes:
      - application/json
      description: 'Perform user login with email and password. The returned `accessToken`
        should be used as a Bearer token in the `Authorization` header (i.e., `Authorization:
        Bearer <accessToken>`) for authenticated endpoints such as Logout.'
      operationId: user-login
      parameters:
      - description: User's credentials
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Authenticates a user using email and password, and returns a token
        pair if successful.
      tags:
      - Account Actions
  /logout:
    post:
      consumes:
      - application/json
      description: Invalidate user session and logout
      operationId: user-logout
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - ApiKeyAuth: []
      summary: Logout user
      tags:
      - Account Actions
  /refresh:
    post:
      consumes:
      - application/json
      description: RefreshToken issues a new access token using a valid refresh token.
      operationId: user-refresh
      parameters:
      - description: Refresh token
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Refresh access token
      tags:
      - Account Actions
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Provide the accessToken as a Bearer token: ''Bearer {accessToken}'''
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"