# handlers/domain_handler.go
package handlers

import (
	"goweb/api"
	"goweb/server"
	"net/http"

	"github.com/brianvoe/gofakeit/v7"
	"github.com/labstack/echo/v4"
)

// DomainHandler provides endpoints for managing domains, including CRUD operations (some not implemented).
type DomainHandler struct {
	server *server.Server
}

// NewDomainHandler initializes the DomainHandler with the provided server.
func NewDomainHandler(server *server.Server) *DomainHandler {
	return &DomainHandler{server: server}
}

// Type returns the string identifier for the DomainHandler.
func (u DomainHandler) Type() string {
	return "Domain"
}

// List godoc
// @Summary List domains
// @Description Returns a sample response for listing domains.
// @ID domain-list
// @Tags Domain Management
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Success 200 {object} api.Response
// @Router /api/domain [get]
func (u DomainHandler) List(e echo.Context) error {
	return api.WebResponse(e, http.StatusOK, api.STATUS_OK(gofakeit.Paragraph(100, 10, 10, " ")))
}

// Create godoc
// @Summary Create domain
// @Description Not implemented.
// @ID domain-create
// @Tags Domain Management
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Failure 404 {object} api.Response
// @Router /api/domain [post]
func (u DomainHandler) Create(e echo.Context) error {
	return api.WebResponse(e, http.StatusNotFound, api.RESOURCE_NOT_FOUND("Create Domain not implemented"))
}

// Read godoc
// @Summary Get domain
// @Description Not implemented.
// @ID domain-read
// @Tags Domain Management
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param uuid path string true "Domain UUID"
// @Failure 404 {object} api.Response
// @Router /api/domain/{uuid} [get]
func (u DomainHandler) Read(e echo.Context) error {
	return api.WebResponse(e, http.StatusNotFound, api.RESOURCE_NOT_FOUND("Read Domain not implemented"))
}

// Update godoc
// @Summary Update domain
// @Description Not implemented.
// @ID domain-update
// @Tags Domain Management
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param uuid path string true "Domain UUID"
// @Failure 404 {object} api.Response
// @Router /api/domain/{uuid} [put]
func (u DomainHandler) Update(e echo.Context) error {
	return api.WebResponse(e, http.StatusNotFound, api.RESOURCE_NOT_FOUND("Update Domain not implemented"))
}

// Delete godoc
// @Summary Delete domain
// @Description Not implemented.
// @ID domain-delete
// @Tags Domain Management
// @Accept json
// @Produce json
// @Security ApiKeyAuth
// @Param uuid path string true "Domain UUID"
// @Failure 404 {object} api.Response
// @Router /api/domain/{uuid} [delete]
func (u DomainHandler) Delete(e echo.Context) error {
	return api.WebResponse(e, http.StatusNotFound, api.RESOURCE_NOT_FOUND("Delete Domain not implemented"))
}